Use FlushedInputStream only when loading from network
